fix(uia): the native handle is unsigned on this road too
I wrote that the Rust road "was already right" when correcting the PowerShell
one. It was not. The UIA property is a VT_I4, so a handle with the high bit
set comes back sign-extended, and rendering it through `isize` produced a
negative decimal string -- which `parseWindowHandle` rejects, so exactly those
windows recorded no handle and kept the behaviour this item exists to end.
Invisibly, because a missing handle is indistinguishable from a build that
cannot report one. The same defect as the PowerShell mask, on the road I had
just declared safe.

Truncated to the 32 bits that are the handle, unsigned. USER handles are
32-bit values sign-extended for interop, so nothing is lost -- and the two
roads now agree, which matters beyond this bug: the same window read through
Rust and through PowerShell yields the same string, and two representations
would have quietly broken merged entities.

`as usize as u32` because HWND is a raw pointer in windows 0.62: the address
first, then the truncation.

Reported by codex on PR #619 (P2).

+++ src/uia/tree.rs
@@ -66,8 +66,22 @@ fn get_elements_impl(ctx: &UiaContext, opts: &GetElementsOptions) -> napi::Resul
     // Zero is filtered rather than reported: `CachedNativeWindowHandle` answers NULL for an element
     // with no host window, and "no window" and "the window numbered 0" must not arrive as the same
     // value — the consumer treats this as the handle to run the coordinate ladder against.
+    //
+    // **`as u32`, not `as isize`.** The UIA property is a VT_I4, so a handle with the high bit set
+    // comes back sign-extended and `isize` renders it as a NEGATIVE decimal string;
+    // `parseWindowHandle` rejects non-positive handles, so exactly those windows would record none
+    // and keep the behaviour this item exists to end — invisibly, since a missing handle is
+    // indistinguishable from a build that cannot report one (PR 側 codex on #619, P2; my own claim
+    // that this road "was already right" was wrong).
+    //
+    // The low 32 bits are the whole handle — USER handles are 32-bit values sign-extended for
+    // interop — so the truncation loses nothing, and it makes the two roads agree: the same window
+    // read through Rust and through PowerShell now yields the same string, which two different
+    // representations would have quietly broken for merged entities.
     let window_hwnd = unsafe { root.CachedNativeWindowHandle().ok() }
-        .map(|h| h.0 as isize)
+        // `as usize as u32`: HWND is a raw pointer in windows 0.62, so the address is taken first
+        // and then truncated to the 32 bits that are the handle.
+        .map(|h| h.0 as usize as u32)
         .filter(|h| *h != 0)
         .map(|h| h.to_string());
